requirements.txt
pytest
mock
PyHamcrest
django
djangorestframework
